+++ CHANGELOG.md
@@ -58,6 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--cli-order` option to generate outputs in arbitrary order. (#106)
 - QR codes generation and update: symbols and footprints. (#93)
 - Support for `--subst-models` option for KiCad 6's kicad2step. (#137)
+- Added global options to define the PCB details (`pcb_material`,
+  `solder_mask_color`, `silk_screen_color` and `pcb_finish`)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current